# Monte-Carlo Engine, seed parameter implementation and new numerical schemes for SDEs