--- a/crates/RustQuant_stochastics/src/arithmetic_brownian_motion.rs
+++ b/crates/RustQuant_stochastics/src/arithmetic_brownian_motion.rs
@@ -58,15 +58,17 @@ impl StochasticProcess for ArithmeticBrownianMotion {
 #[cfg(test)]
 mod tests_abm {
     use super::*;
-    use crate::StochasticProcessConfig;
+    use crate::{StochasticProcessConfig, StochasticScheme};
     use RustQuant_math::*;
     use RustQuant_utils::assert_approx_equal;
 
     #[test]
     fn test_arithmetic_brownian_motion() {
         let abm = ArithmeticBrownianMotion::new(0.05, 0.9);
-        let config = StochasticProcessConfig::new(10.0, 0.0, 0.5, 125, 1000, false);
-        let output = abm.euler_maruyama(&config);
+        let config = StochasticProcessConfig::new(
+            10.0, 0.0, 0.5, 125, StochasticScheme::EulerMaruyama, 1000, false, None
+        );
+        let output = abm.monte_carlo(&config);
 
         // let file1 = "./images/ABM1.png";
         // plot_vector((&output.trajectories[0]).clone(), file1).unwrap();