--- a/crates/RustQuant_stochastics/src/arithmetic_brownian_motion.rs
+++ b/crates/RustQuant_stochastics/src/arithmetic_brownian_motion.rs
@@ -7,7 +7,6 @@
 //      - LICENSE-MIT.md
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-// use crate::models::arithmetic_brownian_motion::ArithmeticBrownianMotion;
 use crate::process::StochasticProcess;
 use crate::ModelParameter;