--- a/crates/RustQuant_stochastics/src/brownian_motion.rs
+++ b/crates/RustQuant_stochastics/src/brownian_motion.rs
@@ -54,7 +54,7 @@ mod sde_tests {
     // use std::time::Instant;
 
     use super::*;
-    use crate::StochasticProcessConfig;
+    use crate::{StochasticProcessConfig, StochasticScheme};
     use RustQuant_math::*;
     use RustQuant_utils::assert_approx_equal;
 
@@ -84,8 +84,10 @@ mod sde_tests {
         // }
         // assert!(1 == 2);
 
-        let config = StochasticProcessConfig::new(0.0, 0.0, 0.5, 100, 1000, false);
-        let output_serial = bm.euler_maruyama(&config);
+        let config = StochasticProcessConfig::new(
+            0.0, 0.0, 0.5, 100, StochasticScheme::EulerMaruyama, 1000, false, None
+        );
+        let output_serial = bm.monte_carlo(&config);
         // let output_parallel = (&bm).euler_maruyama(10.0, 0.0, 0.5, 100, 10, true);
 
         // let file1 = "./images/BM1.png";