--- a/crates/RustQuant_stochastics/src/fractional_cox_ingersoll_ross.rs
+++ b/crates/RustQuant_stochastics/src/fractional_cox_ingersoll_ross.rs
@@ -7,14 +7,9 @@
 //      - LICENSE-MIT.md
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-use super::{
-    fractional_brownian_motion::FractionalProcessGeneratorMethod,
-    process::{StochasticProcess, Trajectories},
-    StochasticProcessConfig,
-};
-use crate::fractional_brownian_motion::FractionalBrownianMotion;
+use crate::process::{StochasticProcessConfig, StochasticProcess, Trajectories};
+use crate::fractional_process::{fractional_monte_carlo, FractionalProcessGeneratorMethod};
 use crate::model_parameter::ModelParameter;
-use rayon::prelude::*;
 
 /// Struct containing the Ornstein-Uhlenbeck process parameters.
 pub struct FractionalCoxIngersollRoss {
@@ -79,47 +74,7 @@ impl StochasticProcess for FractionalCoxIngersollRoss {
     }
 
     fn monte_carlo(&self, config: &StochasticProcessConfig) -> Trajectories {
-        let (t_0, x_0, t_n, n_steps, m_paths, parallel) = config.unpack();
-
-        let fgn = match self.method {
-            FractionalProcessGeneratorMethod::CHOLESKY => {
-                let fbm = FractionalBrownianMotion::new(
-                    self.hurst,
-                    FractionalProcessGeneratorMethod::CHOLESKY,
-                );
-                FractionalBrownianMotion::fgn_cholesky(&fbm, n_steps, t_n)
-            }
-            FractionalProcessGeneratorMethod::FFT => {
-                let fbm = FractionalBrownianMotion::new(
-                    self.hurst,
-                    FractionalProcessGeneratorMethod::FFT,
-                );
-                // FractionalBrownianMotion::fgn_fft(&fbm, n_steps, t_n)
-                fbm.fgn_fft(n_steps, t_n)
-            }
-        };
-
-        let dt: f64 = (t_n - t_0) / (n_steps as f64);
-
-        // Initialise empty paths and fill in the time points.
-        let mut paths = vec![vec![x_0; n_steps + 1]; m_paths];
-        let times: Vec<f64> = (0..=n_steps).map(|t| t_0 + dt * (t as f64)).collect();
-
-        let path_generator = |path: &mut Vec<f64>| {
-            for t in 0..n_steps {
-                path[t + 1] = path[t]
-                    + self.drift(path[t], times[t]) * dt
-                    + self.diffusion(path[t], times[t]) * fgn[t];
-            }
-        };
-
-        if parallel {
-            paths.par_iter_mut().for_each(path_generator);
-        } else {
-            paths.iter_mut().for_each(path_generator);
-        }
-
-        Trajectories { times, paths }
+        fractional_monte_carlo(self, config, &self.method, self.hurst)
     }
 }