@avhz 2024-10-26
Sorry it's taken me a while to have a look at this. 
Just a couple of things:

- Why not put the `StochasticMethod` into the config ?
- Did you benchmark it ?