@yfnaji 2025-03-14
Hey @avhz 

I have moved `StochasticMethod` (now called `StochasticScheme`) into the config + added some benchmark tests.

I have also included other changes that I have noted at the top of the description.