@avhz 2025-03-17
Looks good :) 

Just one thing, I would change the method names to a verb, something like `process.generate()` rather than `process.monte_carlo()`, we can reserve Monte Carlo for pricing and other estimation, rather than path simulation.