@yfnaji 2025-03-17
Thanks for having a look @avhz ! Agreed with your comment, all references to Monte-Carlo (method names, file names and documentation) have been amended 👍